rust: use nickel-helper in cucumber tests
