+++ Cargo.toml
@@ -36,6 +36,8 @@ ron = "0.8"
 serde_json = "1.0"
 futures = "0.3"
 
+smart-keymap-nickel-helper = { path = "./smart-keymap-nickel-helper" }
+
 [build-dependencies]
 smart-keymap-nickel-helper = { path = "./smart-keymap-nickel-helper" }
