+++ tests/cucumber/keymap.rs
@@ -1,6 +1,4 @@
 use std::fmt::Debug;
-use std::io::{self, Write};
-use std::process::{Command, Stdio};
 
 use cucumber::gherkin::Step;
 use cucumber::{given, then, when, World};
@@ -14,82 +12,9 @@ mod common;
 
 use common::Deserializer;
 
-/// Likely reasons why running `nickel` may fail.
-enum NickelError {
-    NickelNotFound,
-    EvalError(String),
-}
-
-/// Result of Nickel evaluation.
-type NickelResult = Result<String, NickelError>;
-
-/// Evaluates the Nickel expr for a keymap, returning the json serialization.
-fn nickel_json_serialization_for_keymap(keymap_ncl: &str) -> NickelResult {
-    let spawn_nickel_result = Command::new("nickel")
-        .args([
-            "export",
-            "--format=json",
-            format!("--import-path={}/ncl", env!("CARGO_MANIFEST_DIR")).as_ref(),
-            "--field=serialized_json_composite_keys",
-        ])
-        .stdin(Stdio::piped())
-        .stderr(Stdio::piped())
-        .stdout(Stdio::piped())
-        .spawn()
-        .map_err(|e| match e.kind() {
-            io::ErrorKind::NotFound => return Err(NickelError::NickelNotFound),
-            _ => panic!("Failed to spawn nickel: {:?}", e),
-        });
-
-    match spawn_nickel_result {
-        Ok(mut nickel_command) => {
-            let child_stdin = nickel_command.stdin.as_mut().unwrap();
-            child_stdin
-                .write_all(
-                    format!(r#"(import "keymap-ncl-to-json.ncl") & ({})"#, keymap_ncl).as_bytes(),
-                )
-                .unwrap_or_else(|e| panic!("Failed to write to stdin: {:?}", e));
-
-            match nickel_command.wait_with_output() {
-                Ok(output) => {
-                    if output.status.success() {
-                        String::from_utf8(output.stdout)
-                            .map_err(|e| panic!("Failed to decode UTF-8: {:?}", e))
-                    } else {
-                        let nickel_error_message = String::from_utf8(output.stderr)
-                            .unwrap_or_else(|e| panic!("Failed to decode UTF-8: {:?}", e));
-                        Err(NickelError::EvalError(nickel_error_message))
-                    }
-                }
-                Err(io_e) => {
-                    panic!("Unhandled IO error: {:?}", io_e)
-                }
-            }
-        }
-        Err(e) => Err(e?),
-    }
-}
-
-/// Evaluates the Nickel expr for an HID, returning the json serialization.
-fn nickel_to_json_for_hid_report(keymap_ncl: &str) -> io::Result<String> {
-    let mut nickel_command = Command::new("nickel")
-        .args([
-            "export",
-            "--format=json",
-            format!("--import-path={}/ncl", env!("CARGO_MANIFEST_DIR")).as_ref(),
-            "--field=as_bytes",
-        ])
-        .stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .spawn()?;
-
-    let child_stdin = nickel_command.stdin.as_mut().unwrap();
-    child_stdin.write_all(format!(r#"(import "hid-report.ncl") & ({})"#, keymap_ncl).as_bytes())?;
-
-    let output = nickel_command.wait_with_output()?;
-
-    String::from_utf8(output.stdout).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-}
+use smart_keymap_nickel_helper::{
+    nickel_json_serialization_for_keymap, nickel_to_json_for_hid_report, NickelError,
+};
 
 type Key = key::composite::Key;
 type Context = key::composite::Context;
@@ -166,7 +91,10 @@ impl Default for KeymapWorld {
 #[given("a keymap.ncl:")]
 fn setup_nickel_keymap(world: &mut KeymapWorld, step: &Step) {
     let keymap_ncl = step.docstring().unwrap();
-    match nickel_json_serialization_for_keymap(keymap_ncl) {
+    match nickel_json_serialization_for_keymap(
+        format!("{}/ncl", env!("CARGO_MANIFEST_DIR")),
+        keymap_ncl,
+    ) {
         Ok(json) => {
             let keys_vec_result: serde_json::Result<Vec<Key>> = serde_json::from_str(&json);
             match keys_vec_result {
@@ -221,7 +149,10 @@ fn when_keymap_tick(world: &mut KeymapWorld, num_ticks: u16) {
 #[then("the HID keyboard report should equal")]
 fn check_report(world: &mut KeymapWorld, step: &Step) {
     let hid_report_ncl = step.docstring().unwrap();
-    match nickel_to_json_for_hid_report(hid_report_ncl) {
+    match nickel_to_json_for_hid_report(
+        format!("{}/ncl", env!("CARGO_MANIFEST_DIR")),
+        hid_report_ncl,
+    ) {
         Ok(json) => {
             let expected_report: Vec<u8> = serde_json::from_str(&json).unwrap();
 
@@ -236,7 +167,10 @@ fn check_report(world: &mut KeymapWorld, step: &Step) {
 #[then("the HID keyboard report from the next tick() should equal")]
 fn check_tick_report(world: &mut KeymapWorld, step: &Step) {
     let hid_report_ncl = step.docstring().unwrap();
-    match nickel_to_json_for_hid_report(hid_report_ncl) {
+    match nickel_to_json_for_hid_report(
+        format!("{}/ncl", env!("CARGO_MANIFEST_DIR")),
+        hid_report_ncl,
+    ) {
         Ok(json) => {
             let expected_report: Vec<u8> = serde_json::from_str(&json).unwrap();
